pyproject.toml
[tool.poetry]
name = "tbmbot"
version = "0.0.1"
description = "TBM bot for Discord"
authors = ["Dysta <xav4098@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
nextcord = {extras = ["speed"], version = "^2.0.0-alpha.8"}
taskipy = "^1.10.1"
python-dotenv = "^0.19.2"
loguru = "^0.6.0"
pydantic = "^1.9.0"
markdownify = "^0.10.3"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
autoflake = "^1.4"

[tool.taskipy.tasks]
start = { cmd = "python -m tbmbot", help = "run the bot" }
black = { cmd = "python -m black .", help = "blakify the code" }
clean = { cmd = "python -m autoflake . -r -i -v --ignore-init-module-imports --remove-all-unused-imports --remove-unused-variables", help = "remove unused code/import/variable" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
